Improved Structure and SubItems Feature Update
