+++ ExpenseTracker.cpp
@@ -1,0 +1,11 @@
+#include<iostream>
+#include"ExpenseTrackerModule.h"
+using namespace std;
+
+
+
+int main() {
+    ExpenseTracker tracker;
+    tracker.menu();
+    return 0;
+}
